Has division freeze bug
Every so often, this commit freezes in the 32-bit division function,
which, if I use "Set PC to cursor", returns to timer1_poll_delay right
after Timer1_Init().  Neither Timer1_Init() or timer1_poll_delay() have
any divsion.  However, the timer1_poll_delay_ms() that calls
timer1_poll_delay() has 32-bit division in its argument list.

+++ adc.h
@@ -12,10 +12,11 @@
 #define SEL_PORTB 1
 #define ADC_NUM_READS 4 // This specifies how many reads the program should average.
 #define FOSC_DIV_8 1
+#include "globaldef.h"
 
-void ADC_Enable(int select, int port_pin);
-unsigned ADC_ReadOne(int channel);
-unsigned ADC_Read(int channel);
+void ADC_Enable(byte select, byte port_pin);
+unsigned ADC_ReadOne(byte channel);
+unsigned ADC_Read(byte channel);
 
 #endif	/* ADC_H */
